Move the version gate tests into their own file
The floor, the pre-floor exact-match fallback and ClassifyPeerVersion answer
one question between them: may these two peers exchange records at all. The
stray "Exact version matching" banner moves down to the describe it labels,
which it had drifted away from.
